major UI change; todo: layers, object_utils, Dialog
