+++ speckle_toolbox/esri/toolboxes/speckle/speckle_arcgis.py
@@ -1,787 +1,565 @@
 
-# -*- coding: utf-8 -*-
-
-from typing import Any, Callable, List, Optional, Tuple
-#r'''
-from collections import defaultdict
-
-import arcpy
-#from arcpy import toolbox
-from arcpy._mp import ArcGISProject, Map, Layer as arcLayer
-from arcpy import metadata as md
-
-try:
-    from speckle.converter.layers.Layer import Layer, RasterLayer
-    from speckle.converter.layers._init_ import convertSelectedLayers, layerToNative, cadLayerToNative, bimLayerToNative
-    from speckle.ui.project_vars import toolboxInputsClass, speckleInputsClass
-    from speckle.converter.layers.emptyLayerTemplates import createGroupLayer
-    from speckle.converter.layers.Layer import VectorLayer
-except: 
-    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.Layer import Layer, RasterLayer
-    from speckle_toolbox.esri.toolboxes.speckle.converter.layers._init_ import convertSelectedLayers, layerToNative, cadLayerToNative, bimLayerToNative
-    from speckle_toolbox.esri.toolboxes.speckle.ui.project_vars import toolboxInputsClass, speckleInputsClass
-    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.emptyLayerTemplates import createGroupLayer
-    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.Layer import VectorLayer
-    
-from arcgis.features import FeatureLayer
-import os
 import os.path
-import sys 
+from typing import Any, Callable, List, Optional, Tuple, Union
 
-import specklepy
-from specklepy.api.models import Branch, Stream, Streams
-from specklepy.transports.server.server import ServerTransport
-from specklepy.api.credentials import get_local_accounts
-from specklepy.api.client import SpeckleClient
+from PyQt5.QtCore import QCoreApplication, QSettings, Qt, QTranslator, QRect 
+from PyQt5.QtGui import QIcon
+from PyQt5.QtWidgets import QAction, QDockWidget, QVBoxLayout, QWidget
+from PyQt5 import QtWidgets
+
 from specklepy.api import operations
-from specklepy.logging.exceptions import (
-    GraphQLException,
-    SpeckleException,
-    SpeckleWarning,
-)
+from specklepy.logging.exceptions import SpeckleException, GraphQLException
 #from specklepy.api.credentials import StreamWrapper
+from specklepy.api.models import Stream
 from specklepy.api.wrapper import StreamWrapper
 from specklepy.objects import Base
-from specklepy.logging import metrics
+from specklepy.transports.server import ServerTransport
+from specklepy.api.credentials import Account, get_local_accounts #, StreamWrapper
+from specklepy.api.client import SpeckleClient
+import webbrowser
 
-#'''
+import arcpy
+from arcpy._mp import ArcGISProject, Map
+from arcpy._mp import Layer as arcLayer
 
-def traverseObject(
-        base: Base,
-        callback: Optional[Callable[[Base], bool]],
-        check: Optional[Callable[[Base], bool]],
+try: 
+    from speckle.plugin_utils.object_utils import callback, traverseObject
+    from speckle.converter.layers.Layer import (Layer, VectorLayer, RasterLayer) 
+    from speckle.converter.layers import convertSelectedLayers, getLayers
+    from speckle.converter.layers.utils import findAndClearLayerGroup
+except: 
+    from speckle_toolbox.esri.toolboxes.speckle.plugin_utils.object_utils import callback, traverseObject
+    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.Layer import (Layer, VectorLayer, RasterLayer)
+    from speckle_toolbox.esri.toolboxes.speckle.converter.layers import convertSelectedLayers, getLayers
+    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.emptyLayerTemplates import createGroupLayer
+    from speckle_toolbox.esri.toolboxes.speckle.converter.layers.utils import findAndClearLayerGroup
+
+from ui.add_stream_modal import AddStreamModalDialog
+from ui.create_stream import CreateStreamModalDialog
+from ui.create_branch import CreateBranchModalDialog
+
+# Import the code for the dialog
+from ui.validation import tryGetStream, validateBranch, validateCommit, validateStream, validateTransport 
+
+SPECKLE_COLOR = (59,130,246)
+SPECKLE_COLOR_LIGHT = (69,140,255)
+
+class SpeckleGIS:
+    """Speckle Connector Plugin for ArcGIS"""
+
+    dockwidget: Optional[QDockWidget]
+    add_stream_modal: AddStreamModalDialog
+    create_stream_modal: CreateStreamModalDialog
+    current_streams: List[Tuple[StreamWrapper, Stream]]  #{id:(sw,st),id2:()}
+    current_layers: List[Tuple[str, arcLayer]] = []
+
+    active_stream: Optional[Tuple[StreamWrapper, Stream]] 
+
+    gis_project: ArcGISProject #QgsProject
+
+    lat: float
+    lon: float
+
+    default_account: Account
+    accounts: List[Account]
+
+    def __init__(self, iface):
+        """Constructor. 
+        """
+        # Save reference to the QGIS interface
+        self.dockwidget = None
+        #self.iface = None
+        self.gis_project = ArcGISProject('CURRENT') #QgsProject.instance()
+        self.current_streams = []
+        self.active_stream = None
+        self.default_account = None 
+        self.accounts = [] 
+
+        self.btnAction = 0
+
+        self.lat = 0.0
+        self.lon = 0.0
+        # initialize plugin directory
+        self.plugin_dir = os.path.dirname(__file__)
+        # initialize locale
+        #locale = QSettings().value("locale/userLocale")[0:2]
+        #locale_path = os.path.join(
+        #    self.plugin_dir, "i18n", "SpeckleQGIS_{}.qm".format(locale)
+        #)
+
+        #if os.path.exists(locale_path):
+        #    self.translator = QTranslator()
+        #    self.translator.load(locale_path)
+        #    QCoreApplication.installTranslator(self.translator)
+
+        # Declare instance attributes
+        self.actions = []
+        self.menu = self.tr("&SpeckleArcGIS")
+
+        # Check if plugin was started the first time in current QGIS session
+        # Must be set in initGui() to survive plugin reloads
+        self.pluginIsActive = False
+
+    # noinspection PyMethodMayBeStatic
+
+    def tr(self, message: str):
+        """Get the translation for a string using Qt translation API.
+
+        We implement this ourselves since we do not inherit QObject.
+
+        :param message: String for translation.
+        :type message: str, QString
+
+        :returns: Translated version of message.
+        :rtype: QString
+        """
+        # noinspection PyTypeChecker,PyArgumentList,PyCallByClass
+        return QCoreApplication.translate("SpeckleGIS", message)
+
+    def add_action(
+        self,
+        icon_path: str,
+        text,
+        callback,
+        enabled_flag=True,
+        add_to_menu=True,
+        add_to_toolbar=True,
+        status_tip=None,
+        whats_this=None,
+        parent=None,
     ):
-        if check and check(base):
-            res = callback(base) if callback else False
-            if res:
-                return
-        memberNames = base.get_member_names()
-        for name in memberNames:
-            try:
-                if ["id", "applicationId", "units", "speckle_type"].index(name):
-                    continue
-            except:
-                pass
-            traverseValue(base[name], callback, check)
-            
-def traverseValue(
-    value: Any,
-    callback: Optional[Callable[[Base], bool]],
-    check: Optional[Callable[[Base], bool]],
-):
-    if isinstance(value, Base):
-        traverseObject(value, callback, check)
-    if isinstance(value, List):
-        for item in value:
-            traverseValue(item, callback, check)
+        """Add a toolbar icon to the toolbar.
 
+        :param icon_path: Path to the icon for this action. Can be a resource
+            path (e.g. ':/plugins/foo/bar.png') or a normal file system path.
+        :type icon_path: str
 
-class Toolbox:
-    def __init__(self):
-        """Define the toolbox (the name of the toolbox is the name of the
-        .pyt file)."""
-        print("___ping_Toolbox")
-        self.label = "Speckle Tools"
-        self.alias = "speckle_toolbox_"  
-        # List of tool classes associated with this toolbox
-        self.tools = [Speckle]  
+        :param text: Text that should be shown in menu items for this action.
+        :type text: str
 
-        try: 
-            version = arcpy.GetInstallInfo()['Version']
-            python_version = f"python {'.'.join(map(str, sys.version_info[:2]))}"
-            metrics.set_host_app("ArcGIS", ', '.join([f"ArcGIS {version}", python_version])) 
-        except: 
-            metrics.set_host_app("ArcGIS")
-        # https://pro.arcgis.com/en/pro-app/2.8/arcpy/mapping/alphabeticallistofclasses.htm#except: print("something happened")
+        :param callback: Function to be called when the action is triggered.
+        :type callback: function
 
-class Speckle:
-    def __init__(self):  
-        print("__________________INIT SPECKLE TOOL_________")
-        self.label       = "Speckle"
-        self.description = "Allows you to send and receive your layers " + \
-                           "to/from other software using Speckle server." 
+        :param enabled_flag: A flag indicating if the action should be enabled
+            by default. Defaults to True.
+        :type enabled_flag: bool
 
-        self.toRefresh = False
-        self.speckleInputs = None
-        self.toolboxInputs = None
+        :param add_to_menu: Flag indicating whether the action should also
+            be added to the menu. Defaults to True.
+        :type add_to_menu: bool
 
-        total = len(speckleInputsClass.instances)
-        print(total)
-        for i in range(total):
-            if speckleInputsClass.instances[total-i-1] is not None: 
-                try: 
-                    y = speckleInputsClass.instances[total-i-1].streams_default 
-                    #if not isinstance(speckleInputsClass.instances[total-i-1].streams_default, SpeckleException): # also will throw exception in case not initialized properly 
-                    self.speckleInputs = speckleInputsClass.instances[total-i-1] # take latest (first in reverted list)
-                    break
-                except: pass
-        if self.speckleInputs is None or isinstance(self.speckleInputs.streams_default, SpeckleException): self.speckleInputs = speckleInputsClass()
-        #print(self.speckleInputs.streams_default)
-        print(len(speckleInputsClass.instances))
+        :param add_to_toolbar: Flag indicating whether the action should also
+            be added to the toolbar. Defaults to True.
+        :type add_to_toolbar: bool
 
-        total = len(toolboxInputsClass.instances)
+        :param status_tip: Optional text to show in a popup when mouse pointer
+            hovers over the action.
+        :type status_tip: str
 
-        for i in range(total):
-            if toolboxInputsClass.instances[total-i-1] is not None: 
-                self.toolboxInputs = toolboxInputsClass.instances[total-i-1] # take latest (first in reverted list)
-                break
-        if self.toolboxInputs is None: self.toolboxInputs = toolboxInputsClass()
+        :param parent: Parent widget for the new action. Defaults None.
+        :type parent: QWidget
 
-        #print(self.speckleInputs.accounts)
-        if len(self.speckleInputs.accounts) == 0:
-            arcpy.AddError("Speckle accounts not found")
+        :param whats_this: Optional text to show in the status bar when the
+            mouse pointer hovers over the action.
 
-    def getParameterInfo(self):
-        #data types: https://pro.arcgis.com/en/pro-app/2.8/arcpy/geoprocessing_and_python/defining-parameter-data-types-in-a-python-toolbox.htm
-        # parameter details: https://pro.arcgis.com/en/pro-app/latest/arcpy/geoprocessing_and_python/customizing-tool-behavior-in-a-python-toolbox.htm
-        print("Get parameter values")
-        cat1 = "Add Streams"
-        cat2 = "Send/Receive"
-        cat3 = "Create custom Spatial Reference"
+        :returns: The action that was created. Note that the action is also
+            added to self.actions list.
+        :rtype: QAction
+        """
 
-        streamsDefalut = arcpy.Parameter(
-            displayName="Add stream from default account",
-            name="streamsDefalut",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            category=cat1
-            )
-        streamsDefalut.filter.type = 'ValueList'
-        if isinstance(self.speckleInputs.streams_default, SpeckleException):
-            arcpy.AddError("Speckle account not accessible")
-            streamsDefalut.filter.list = []
-        elif self.speckleInputs.streams_default is not None:
-            streamsDefalut.filter.list = [ (str(st.name) + " - " + str(st.id)) for st in self.speckleInputs.streams_default ]
-        else:
-            streamsDefalut.filter.list = []
-            arcpy.AddError("Error connecting to default Speckle account")
-        
-        addDefStreams = arcpy.Parameter(
-            displayName="Add",
-            name="addDefStreams",
-            datatype="GPBoolean",
-            parameterType="Optional",
-            direction="Input",
-            category=cat1
-            )
-        addDefStreams.value = False 
+        icon = QIcon(icon_path)
+        action = QAction(icon, text, parent)
+        action.triggered.connect(callback)
+        action.setEnabled(enabled_flag)
 
-        streamUrl = arcpy.Parameter(
-            displayName="Add stream by URL",
-            name="streamUrl",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            category=cat1
-            )
-        streamUrl.value = ""
+        if status_tip is not None:
+            action.setStatusTip(status_tip)
 
-        addUrlStreams = arcpy.Parameter(
-            displayName="Add",
-            name="addUrlStreams",
-            datatype="GPBoolean",
-            parameterType="Optional",
-            direction="Input",
-            category=cat1
-            )
-        addUrlStreams.value = False 
+        if whats_this is not None:
+            action.setWhatsThis(whats_this)
 
-        ############################################################################
-        
-        lat = arcpy.Parameter(
-            displayName="Origin point LAT",
-            name="lat",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            category=cat3
-            )
-        lat.value = str(self.toolboxInputs.lat)
+        #if add_to_toolbar:
+        #    # Adds plugin icon to Plugins toolbar
+        #    self.iface.addToolBarIcon(action)
 
-        lon = arcpy.Parameter(
-            displayName="Origin point LON",
-            name="lon",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            category=cat3
-            )
-        lon.value = str(self.toolboxInputs.lon)
-        
-        setLatLon = arcpy.Parameter(
-            displayName="Create and apply",
-            name="setLatLon",
-            datatype="GPBoolean",
-            parameterType="Optional",
-            direction="Input",
-            category=cat3
-            )
-        setLatLon.value = False
-        
-        ####################################################################################################
+        #if add_to_menu and self.menu:
+        #    self.iface.addPluginToWebMenu(self.menu, action)
 
-        savedStreams = arcpy.Parameter(
-            displayName="Select Stream",
-            name="savedStreams",
-            datatype="GPString",
-            parameterType="Required",
-            direction="Input",
-            multiValue=False,
-            )
-        savedStreams.filter.list = [f"Stream not accessible - {stream[0].stream_id}" if stream[1] is None or isinstance(stream[1], SpeckleException) else f"{stream[1].name} - {stream[1].id}" for i,stream in enumerate(self.speckleInputs.saved_streams)] 
+        self.actions.append(action)
 
-        removeStream = arcpy.Parameter(
-            displayName="Remove",
-            name="removeStream",
-            datatype="GPBoolean",
-            parameterType="Optional",
-            direction="Input"
-            )
-        removeStream.value = False 
+        return action
 
-        branch = arcpy.Parameter(
-            displayName="Branch",
-            name="branch",
-            datatype="GPString",
-            parameterType="Required",
-            direction="Input",
-            )
-        branch.value = ""
-        branch.filter.type = 'ValueList'
+    def initGui(self):
+        """Create the menu entries and toolbar icons inside the QGIS GUI."""
 
-        commit = arcpy.Parameter(
-            displayName="Commit",
-            name="commit",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            )
-        commit.value = ""
-        commit.filter.type = 'ValueList'
+        icon_path = "" #":/plugins/speckle_qgis/icon.png"
+        self.add_action(
+            icon_path,
+            text=self.tr("SpeckleQGIS"),
+            callback=self.run,
+            add_to_menu=False,
+            add_to_toolbar=False,
+            parent=None, #self.iface.mainWindow(),
+        )
 
-        msg = arcpy.Parameter(
-            displayName="Message",
-            name="msg",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            multiValue=False,
-            )
-        msg.value = ""
-
-        selectedLayers = arcpy.Parameter(
-            displayName="Selected Layers",
-            name="selectedLayers",
-            datatype="GPString",
-            parameterType="Optional",
-            direction="Input",
-            multiValue=True,
-            )
-        selectedLayers.filter.list = [str(i) + "-" + l.longName for i,l in enumerate(self.speckleInputs.all_layers)] #"Polyline"
+    def onClosePlugin(self):
+        """Cleanup necessary items here when plugin dockwidget is closed"""
 
         
-        action = arcpy.Parameter(
-            displayName="",
-            name="action",
-            datatype="GPString",
-            parameterType="Required",
-            direction="Input",
-            multiValue=False,
-            )
-        action.value = "Send" 
-        action.filter.list = ["Send", "Receive"]  
 
-        refresh = arcpy.Parameter(
-            displayName="Refresh",
-            name="refresh",
-            datatype="GPBoolean",
-            parameterType="Optional",
-            direction="Input"
-            ) 
-        refresh.value = False 
+        # disconnects
+        if self.dockwidget:
+            try: 
+                self.dockwidget.closingPlugin.disconnect(self.onClosePlugin)
+                self.dockwidget.close()
+            except: pass 
 
-        parameters = [streamsDefalut, addDefStreams, streamUrl, addUrlStreams, lat, lon, setLatLon, savedStreams, removeStream, branch, commit, selectedLayers, msg, action, refresh]
-        return parameters
+        self.pluginIsActive = False
+        # remove this statement if dockwidget is to remain
+        # for reuse if plugin is reopened
+        # Commented next statement since it causes QGIS crashe
+        # when closing the docked window:
+        # self.dockwidget = None
 
-    def isLicensed(self): #optional
-        return True
-
-    def updateParameters(self, parameters: List, toRefresh = False): #optional
-        print("UPDATING PARAMETERS")
         
-        for i, par in enumerate(parameters): 
-            
-
-            if par.name == "addDefStreams" and par.altered and par.value == True: 
-                for p in parameters:
-                    if p.name == "streamsDefalut" and p.valueAsText is not None:
-
-                        # add value from streamsDefault to saved streams
-                        selected_stream_name = p.valueAsText[:]
-
-                        for stream in self.speckleInputs.streams_default:
-                            if stream.name == selected_stream_name.split(" - ")[0]:
-                                print("_____Add from list___")
-                                wr = StreamWrapper(f"{self.speckleInputs.account.serverInfo.url}/streams/{stream.id}?u={self.speckleInputs.account.userInfo.id}")
-                                self.toolboxInputs.setProjectStreams(wr)
-                                
-                                for p_saved in parameters:
-                                    if p_saved.name == "savedStreams": 
-                                        saved_streams = self.speckleInputs.getProjectStreams()
-                                        self.speckleInputs.saved_streams = saved_streams
-                                        p_saved.filter.list = [f"Stream not accessible - {stream[0].stream_id}" if stream[1] is None or isinstance(stream[1], SpeckleException) else f"{stream[1].name} - {stream[1].id}" for i,stream in enumerate(saved_streams)] 
-                                        if len(p_saved.filter.list)>0: print(p_saved.filter.list); p_saved.value = p_saved.filter.list[0]
-                                break
-                        p.value = None
-                par.value = False
-            
-            if par.name == "addUrlStreams" and par.altered and par.value == True: 
-                for p in parameters:
-                    if p.name == "streamUrl" and p.valueAsText is not None:
-
-                        # add value from streamsDefault to saved streams
-                        query = p.valueAsText[:]
-                        if "http" in query and len(query.split("/")) >= 3: # URL
-                            steamId = query
-                            try: steamId = query.split("/streams/")[1].split("/")[0] 
-                            except: pass
-
-                            # query stream, add to saved
-                            stream = self.speckleInputs.speckle_client.stream.get(id = steamId, branch_limit = 100, commit_limit = 100)
-                            if isinstance(stream, Stream): 
-                                print("_____Add by URL___")
-                                wr = StreamWrapper(f"{self.speckleInputs.account.serverInfo.url}/streams/{stream.id}?u={self.speckleInputs.account.userInfo.id}")
-                                self.toolboxInputs.setProjectStreams(wr)
-                                
-                                for p_saved in parameters:
-                                    if p_saved.name == "savedStreams": 
-                                        saved_streams = self.speckleInputs.getProjectStreams()
-                                        self.speckleInputs.saved_streams = saved_streams
-                                        p_saved.filter.list = [f"Stream not accessible - {st[0].stream_id}" if st[1] is None or isinstance(st[1], SpeckleException) else f"{st[1].name} - {st[1].id}" for i,st in enumerate(saved_streams)] 
-                                        if len(p_saved.filter.list)>0: print(p_saved.filter.list); p_saved.value = p_saved.filter.list[0]
-                            else: pass
-
-                        p.value = None
-                        break
-                par.value = False
-
-
-            if par.name == "removeStream" and par.altered and par.value == True: 
-                for p in parameters:
-                    if p.name == "savedStreams" and p.valueAsText is not None:
-
-                        # get value from savedStreams 
-                        selected_stream_name = p.valueAsText[:]
-                        for streamTup in self.speckleInputs.saved_streams:
-                            stream = streamTup[1]
-                            if stream.name == selected_stream_name.split(" - ")[0]:
-                                print("_____Remove stream___")
-                                wr = StreamWrapper(f"{self.speckleInputs.account.serverInfo.url}/streams/{stream.id}?u={self.speckleInputs.account.userInfo.id}")
-                                self.toolboxInputs.setProjectStreams(wr, False)
-                                
-                                for p_saved in parameters:
-                                    if p_saved.name == "savedStreams": 
-                                        saved_streams = self.speckleInputs.getProjectStreams()
-                                        self.speckleInputs.saved_streams = saved_streams
-                                        p_saved.filter.list = [f"Stream not accessible - {st[0].stream_id}" if st[1] is None or isinstance(st[1], SpeckleException) else f"{st[1].name} - {st[1].id}" for i,st in enumerate(saved_streams)] 
-                                        p_saved.value = None
-                                break
-                        p.value = None
-                par.value = False
-
-            ####################################################################### 
-            if par.name == "setLatLon" and par.altered and par.value == True: 
-                lat = lon = 0
-                for p in parameters:
-                    if p.name == "lat" and p.valueAsText is not None:
-                        # add value from the UI to saved lat
-                        lat = p.valueAsText[:].replace(",","").replace(" ","").replace(";","").replace("_","")
-                        try: lat = float(lat)
-                        except: lat = 0; p.value = "0.0"
-
-                    if p.name == "lon" and p.valueAsText is not None:
-                        # add value from the UI to saved lat
-                        lon = p.valueAsText[:].replace(",","").replace(" ","").replace(";","").replace("_","")
-                        try: lon = float(lon)
-                        except: lon = 0; p.value = "0.0"
-                    coords = [lat, lon]
-                    self.toolboxInputs.set_survey_point(coords)
-                par.value = False
-            
-            ####################################################################### 
-
-            if par.name == "savedStreams" and par.altered:
-                # Search for the stream by name
-                if par.value is not None and "Stream not accessible" not in par.valueAsText[:]:
-                    selected_stream_name = par.valueAsText[:]
-                    self.toolboxInputs.active_stream = None
-                    self.toolboxInputs.active_stream_wrapper = None
-                    for st in self.speckleInputs.saved_streams:
-                        if st[1].name == selected_stream_name.split(" - ")[0]: 
-                            self.toolboxInputs.active_stream = st[1]
-                            self.toolboxInputs.active_stream_wrapper = st[0]
-                            break
-
-                    # edit branches: globals and UI 
-                    branch_list = [branch.name for branch in self.toolboxInputs.active_stream.branches.items]
-                    for p in parameters:
-                        if p.name == "branch":
-                            p.filter.list = branch_list
-
-                            if p.valueAsText not in branch_list: 
-                                p.value = "main"
-                            for b in self.toolboxInputs.active_stream.branches.items:
-                                if b.name == p.value: 
-                                    self.toolboxInputs.active_branch = b
-                                    break 
-                    
-                    # setting commit value and list 
-                    for p in parameters:
-                        if p.name == "commit":
-                            try: 
-                                p.filter.list = [f"{commit.id}"+ " - " + f"{commit.message}" for commit in self.toolboxInputs.active_branch.commits.items]
-                                if p.valueAsText not in p.filter.list:
-                                    p.value = self.toolboxInputs.active_branch.commits.items[0].id + " - " + self.toolboxInputs.active_branch.commits.items[0].message 
-                                    self.toolboxInputs.active_commit = self.toolboxInputs.active_branch.commits.items[0] 
-                            except: 
-                                p.filter.list = []
-                                p.value = None
-                                self.toolboxInputs.active_commit = None 
-                else: par.value = None
-
-            if par.name == "branch" and par.altered: # branches
-                if par.value is not None:
-                    selected_branch_name = par.valueAsText[:]
-                    self.toolboxInputs.active_branch = None
-                    if self.toolboxInputs.active_stream is not None:
-                        for br in self.toolboxInputs.active_stream.branches.items:
-                            if br.name == selected_branch_name:
-                                self.toolboxInputs.active_branch = br
-                                break
-                # edit commit values 
-                if self.toolboxInputs.active_branch is not None: 
-                    for p in parameters:
-                        if p.name == "commit":
-                            try: 
-                                p.filter.list = [f"{commit.id}"+ " - " + f"{commit.message}" for commit in self.toolboxInputs.active_branch.commits.items]
-                                if p.valueAsText not in p.filter.list:
-                                    p.value = self.toolboxInputs.active_branch.commits.items[0].id + " - " + self.toolboxInputs.active_branch.commits.items[0].message 
-                                    self.toolboxInputs.active_commit = self.toolboxInputs.active_branch.commits.items[0]
-                            except: 
-                                p.filter.list = []
-                                p.value = None
-                                self.toolboxInputs.active_commit = None 
-
-            if par.name == "commit" and par.altered: # commits
-                if par.value is not None:
-                    selected_commit_id = par.valueAsText[:].split(" - ")[0]
-                    self.toolboxInputs.active_commit = None
-                    if self.toolboxInputs.active_branch is not None:
-                        for c in self.toolboxInputs.active_branch.commits.items:
-                            if c.id == selected_commit_id: 
-                                self.toolboxInputs.active_commit = c 
-                                break
-            
-            if par.name == "selectedLayers" and par.altered: # selected layers
-                if par.value is not None:
-                    self.toolboxInputs.selected_layers = par.values
-
-            if par.name == "msg" and par.altered and par.valueAsText is not None:
-                self.toolboxInputs.messageSpeckle = par.valueAsText
-                print(self.toolboxInputs.messageSpeckle)
-
-            if par.name == "action" and par.altered:
-                if par.valueAsText == "Send": self.toolboxInputs.action = 1
-                else: self.toolboxInputs.action = 0
-
-            if par.name == "refresh" and par.altered: # refresh btn
-                if par.value == True: 
-                    self.refresh(parameters) 
-            if self.toRefresh == True:
-                self.refresh(parameters) 
-                self.toRefresh = False
         
-        print("____________________________parameters___________________________")
-        return 
-    
-    def refresh(self, parameters: List[Any]): 
-        print("Refresh______")
-        self.speckleInputs: speckleInputsClass = speckleInputsClass()
-        self.toolboxInputs: toolboxInputsClass = toolboxInputsClass()
-        
-        for par in parameters: 
-            if par.name == "streamUrl": par.value = None
-            if par.name == "streamsDefalut": par.value = None
-            if par.name == "savedStreams": par.value = None
-            if par.name == "branch": par.value = ""; par.filter.list = []
-            if par.name == "commit": par.value = None; par.filter.list = []
-            if par.name == "selectedLayers": par.value = None
-            if par.name == "msg": par.value = ""
-            if par.name == "action": par.value = "Send"
-            if par.name == "refresh": par.value = False
-            
-            if par.name == "lat": par.value = str(self.toolboxInputs.get_survey_point()[0])
-            if par.name == "lon": par.value = str(self.toolboxInputs.get_survey_point()[1])
-            if par.name == "streamsDefalut": 
-                if isinstance(self.speckleInputs.streams_default, SpeckleException):
-                    arcpy.AddError("Speckle account not accessible")
-                    par.filter.list = []
-                else:
-                    par.filter.list = [ (st.name + " - " + st.id) for st in self.speckleInputs.streams_default ]
-            if par.name == "savedStreams": 
-                saved_streams = self.speckleInputs.getProjectStreams()
-                par.filter.list = [f"Stream not accessible - {stream[0].stream_id}" if stream[1] is None or isinstance(stream[1], SpeckleException) else f"{stream[1].name} - {stream[1].id}" for i,stream in enumerate(saved_streams)] 
-            if par.name == "selectedLayers": par.filter.list = [str(i) + "-" + l.longName for i,l in enumerate(self.speckleInputs.all_layers)]
-        
-        return parameters
 
-    def updateMessages(self, parameters): #optional
+    def unload(self):
+        """Removes the plugin menu item and icon from QGIS GUI."""
         return
+        #for action in self.actions:
+        #    self.iface.removePluginWebMenu(self.tr("&SpeckleQGIS"), action)
+        #    self.iface.removeToolBarIcon(action)
 
-    def execute(self, parameters: List, messages): 
-        # https://pro.arcgis.com/en/pro-app/latest/arcpy/get-started/what-is-arcpy-.htm
-        #Warning if any of the fields is invalid/empty 
-        print("___________________________Run___________________________")
-        check = self.validateStreamBranch(parameters) # apparently pdate needed to assign proper self.values
+    def onRunButtonClicked(self):
+        if self.btnAction == 0: self.onSend()
+        elif self.btnAction == 1: self.onReceive()
+
+    def onSend(self):
+        """Handles action when Send button is pressed."""
+        if not self.dockwidget: return
+
+        # creating our parent base object
+        project = self.gis_project
+        projectCRS = project.crs()
+        #layerTreeRoot = project.layerTreeRoot()
+
+        bySelection = True
+        if self.dockwidget.layerSendModeDropdown.currentIndex() == 1: bySelection = False 
+        layers = getLayers(self, bySelection) # List[QgsLayerTreeNode]
+
+        # Check if stream id/url is empty
+        if self.active_stream is None:
+            arcpy.AddWarning("Please select a stream from the list.")
+            return
         
-        print(self.toolboxInputs.selected_layers)
-        print(self.toolboxInputs.action)
-        
-        if self.toolboxInputs.action == 1 and check is True: self.onSend(parameters)
-        elif self.toolboxInputs.action == 0 and check is True: self.onReceive(parameters)
-        print("__________________________Run_end___________________________")
-
-    def validateStreamBranch(self, parameters: List):
-        	
-        self.updateParameters(parameters)
-        if self.toolboxInputs.active_stream is None:
-            arcpy.AddError("Choose a valid stream")
-            return False
-        if self.toolboxInputs.active_branch is None: 
-            arcpy.AddError("Choose a valid branch")
-            return False
-        return True 
-
-    def onSend(self, parameters: List):
-
-        print("______________SEND_______________")
-
-        if len(self.toolboxInputs.selected_layers) == 0: 
-            arcpy.AddError("No layers selected for sending")
+        # Check if no layers are selected
+        if len(layers) == 0: #len(selectedLayerNames) == 0:
+            arcpy.AddWarning("No layers selected")
             return
 
-        streamId = self.toolboxInputs.active_stream.id #stream_id
-        client = self.toolboxInputs.active_stream_wrapper.get_client()
-        #client = self.speckleInputs.speckle_client # ?
-        
-        # Get the stream wrapper
-        #streamWrapper = StreamWrapper(None)
-        #client = streamWrapper.get_client()
-        # Ensure the stream actually exists
-        #try:
-        #    client.stream.get(streamId)
-        #except SpeckleException as error:
-        #    print(str(error))
-        #    return
-
-        # next create a server transport - this is the vehicle through which you will send and receive
-        transport = ServerTransport(client=client, stream_id=streamId)
-
-        ##################################### conversions ################################################
         base_obj = Base(units = "m")
-        base_obj.layers = convertSelectedLayers(self.speckleInputs.all_layers, self.toolboxInputs.selected_layers, self.speckleInputs.project)
-        
-        if len(base_obj.layers) == 0: 
-            arcpy.AddMessage("No data sent to stream " + streamId)
+        base_obj.layers = convertSelectedLayers(layers, [],[], projectCRS, project)
+        if base_obj.layers is None:
             return 
+
+        # Reset Survey point
+        self.dockwidget.populateSurveyPoint(self)
+
+        # Get the stream wrapper
+        streamWrapper = self.active_stream[0]
+        streamName = self.active_stream[1].name
+        streamId = streamWrapper.stream_id
+        client = streamWrapper.get_client()
+
+        stream = validateStream(streamWrapper)
+        if stream == None: return
+        
+        branchName = str(self.dockwidget.streamBranchDropdown.currentText())
+        branch = validateBranch(stream, branchName, False)
+        if branch == None: return
+
+        transport = validateTransport(client, streamId)
+        if transport == None: return
         try:
             # this serialises the block and sends it to the transport
             objId = operations.send(base=base_obj, transports=[transport])
-        except SpeckleException as error:
-            arcpy.AddError("Error sending data")
-            #print("Error sending data")
+        except SpeckleException as e:
+            arcpy.AddError("Error sending data: " + str(e.message))
             return
-        except SpeckleWarning as warning:
-            arcpy.AddMessage("SpeckleWarning: " + str(warning.args[0]))
-            
 
-        message = self.toolboxInputs.messageSpeckle
-        print(message)
-        if message is None or ( isinstance(message, str) and len(message) == 0):  message = "Sent from ArcGIS"
-        print(message)
+        message = str(self.dockwidget.messageInput.text())
         try:
             # you can now create a commit on your stream with this object
-            client.commit.create(
+            commit_id = client.commit.create(
                 stream_id=streamId,
                 object_id=objId,
-                branch_name=self.toolboxInputs.active_branch.name,
-                message=message,
-                source_application="ArcGIS",
+                branch_name=branchName,
+                message="Sent objects from QGIS" if len(message) == 0 else message,
+                source_application="QGIS",
             )
             arcpy.AddMessage("Successfully sent data to stream: " + streamId)
-        except:
-            arcpy.AddError("Error creating commit")
+            self.dockwidget.messageInput.setText("")
+            url = streamWrapper.stream_url.split("?")[0] + "/commits/" + commit_id
 
-    def onReceive(self, parameters: List[Any]): 
-        
-        print("______________RECEIVE_______________")
+            # create a temporary floating button 
+            width = self.dockwidget.frameSize().width()
+            height = self.dockwidget.frameSize().height()
+            backgr_color = f"background-color: rgb{str(SPECKLE_COLOR)};"
+            backgr_color_light = f"background-color: rgb{str(SPECKLE_COLOR_LIGHT)};"
+            commit_link_btn = QtWidgets.QPushButton(f"👌 Data sent \n Sent to '{streamName}', view it online")
+            commit_link_btn.setStyleSheet("QPushButton {color: white;border: 0px;border-radius: 17px;padding: 20px;height: 40px;text-align: left;"+ f"{backgr_color}" + "} QPushButton:hover { "+ f"{backgr_color_light}" + " }")
 
-        #if self.validateStreamBranch(parameters) == False: return
-
-        try: 
-            streamId = self.toolboxInputs.active_stream.id #stream_id
-            client = self.toolboxInputs.active_stream_wrapper.get_client()
-            #client = self.speckleInputs.speckle_client # 
-        except SpeckleWarning as warning: 
-            arcpy.AddWarning(str(warning.args[0]))
-
-        # get commit 
-        commit = None 
-        try: 
-            #commit = self.toolboxInputs.active_branch.commits.items[0]
-            commit = self.toolboxInputs.active_commit
-            commitId = commit.id # text to make sure commit exists 
-        except: 
-            try: 
-                commit = self.toolboxInputs.active_branch.commits.items[0]
-                commitId = commit.id 
-                arcpy.AddWarning("Failed to find a commit. Getting the last commit of the branch")
-            except:
-                arcpy.AddError("Failed to find a commit")
-                return
-
-        # next create a server transport - this is the vehicle through which you will send and receive
-        try: 
-            transport = ServerTransport(client=client, stream_id=streamId)
+            widget = QWidget()
+            widget.setAccessibleName("commit_link")
+            connect_box = QVBoxLayout(widget)
+            connect_box.addWidget(commit_link_btn) #, alignment=Qt.AlignCenter) 
+            connect_box.setContentsMargins(0, 0, 0, 0)
+            connect_box.setAlignment(Qt.AlignBottom)  
+            widget.setGeometry(0, 0, width, height)
+            widget.mouseReleaseEvent = lambda event: self.closeWidget()
+            self.dockwidget.link = widget 
             
+            self.dockwidget.layout().addWidget(widget)
+            commit_link_btn.clicked.connect(lambda: self.openLink(url))
+
+        except SpeckleException as e:
+            arcpy.AddError("Error creating commit")
+    
+    def openLink(self, url):
+        webbrowser.open(url, new=0, autoraise=True)
+        self.closeWidget()
+
+    def closeWidget(self):
+        # https://stackoverflow.com/questions/5899826/pyqt-how-to-remove-a-widget
+        return 
+        #self.dockwidget.layout().removeWidget(self.dockwidget.link)
+        #sip.delete(self.dockwidget.link)
+        #self.dockwidget.link = None
+
+    def onReceive(self):
+        """Handles action when the Receive button is pressed"""
+
+        if not self.dockwidget: return
+
+        # Check if stream id/url is empty
+        if self.active_stream is None:
+            arcpy.AddWarning("Please select a stream from the list.")
+            return
+
+        # Get the stream wrapper
+        streamWrapper = self.active_stream[0]
+        streamId = streamWrapper.stream_id
+        client = streamWrapper.get_client()
+        # Ensure the stream actually exists
+        try:
+            stream = validateStream(streamWrapper)
+            if stream == None: return
+            
+            branchName = str(self.dockwidget.streamBranchDropdown.currentText())
+            branch = validateBranch(stream, branchName, True)
+            if branch == None: return
+
+            commitId = str(self.dockwidget.commitDropdown.currentText())
+            commit = validateCommit(branch, commitId)
+            if commit == None: return
+
+        except SpeckleException as error:
+            arcpy.AddError(str(error))
+            return
+
+        transport = validateTransport(client, streamId)
+        if transport == None: return 
+        
+        try:
+            objId = commit.referencedObject
+            #commitDetailed = client.commit.get(streamId, commit.id)
+            app = commit.sourceApplication
+            if branch.name is None or commit.id is None or objId is None: return 
+
+            commitObj = operations.receive(objId, transport, None)
+
             client.commit.received(
             streamId,
             commit.id,
-            source_application="ArcGIS",
-            message="Received commit in ArcGIS",
+            source_application="QGIS",
+            message="Received commit in QGIS",
             )
-        except: 
-            arcpy.AddError("Make sure your account has access to the chosen stream")
-            return
-        try:
-            #print(commit)
-            objId = commit.referencedObject
-            commitDetailed = client.commit.get(streamId, commit.id)
-            if isinstance(commitDetailed, GraphQLException): 
-                arcpy.AddError("Access error")
-                return
-            app = commitDetailed.sourceApplication
-            if objId is None:
-                return
-            commitObj = operations.receive(objId, transport, None)
-            
+
             if app != "QGIS" and app != "ArcGIS": 
-                if self.speckleInputs.project.activeMap.spatialReference.type == "Geographic" or self.speckleInputs.project.activeMap.spatialReference is None: #TODO test with invalid CRS
+                if self.gis_project.activeMap.spatialReference.type == "Geographic" or self.gis_project.activeMap.spatialReference is None: #TODO test with invalid CRS
                     arcpy.AddMessage("It is advisable to set the project Spatial reference to Projected type before receiving CAD geometry (e.g. EPSG:32631), or create a custom one from geographic coordinates")
                     print("It is advisable to set the project Spatial reference to Projected type before receiving CAD geometry (e.g. EPSG:32631), or create a custom one from geographic coordinates")
-            print(f"Successfully received {objId}")
+            arcpy.AddMessage(f"Succesfully received {objId}")
 
-            # Clear 'latest' group
-            streamBranch = streamId + "_" + self.toolboxInputs.active_branch.name + "_" + str(commit.id)
-            streamBranch = streamBranch.replace("[","_").replace("]","_").replace(" ","_").replace("-","_").replace("(","_").replace(")","_").replace(":","_").replace("\\","_").replace("/","_").replace("\"","_").replace("&","_").replace("@","_").replace("$","_").replace("%","_").replace("^","_")
-    
-            newGroupName = f'{streamBranch}'
+            # If group exists, remove layers inside  
+            newGroupName = streamId + "_" + branch.name + "_" + commit.id
             
-            groupExists = 0
-            print(newGroupName)
-            for l in self.speckleInputs.project.activeMap.listLayers(): 
-                #print(l.longName)
-                if l.longName.startswith(newGroupName + "\\"):
-                    #print(l.longName)
-                    self.speckleInputs.project.activeMap.removeLayer(l)
-                    groupExists+=1
-                elif l.longName == newGroupName: 
-                    groupExists+=1
-            print(newGroupName)
-            if groupExists == 0:
-                # create empty group layer file 
-                path = self.speckleInputs.project.filePath.replace("aprx","gdb") #"\\".join(self.toolboxInputs.project.filePath.split("\\")[:-1]) + "\\speckle_layers\\"
-                print(path)
-                try:
-                    f = open(path + "\\" + newGroupName + ".lyrx", "w")
-                    content = createGroupLayer().replace("TestGroupLayer", newGroupName)
-                    f.write(content)
-                    f.close()
-                    newGroupLayer = arcpy.mp.LayerFile(path + "\\" + newGroupName + ".lyrx")
-                    layerGroup = self.speckleInputs.project.activeMap.addLayer(newGroupLayer)[0]
-                except: # for 3.0.0
-                    if self.speckleInputs.active_map is not None:
-                        layerGroup = self.speckleInputs.active_map.createGroupLayer(newGroupName)
-                    else:
-                        arcpy.AddWarning("The map didn't fully load, try refreshing the plugin.")
-                        return
+            findAndClearLayerGroup(self.gis_project, newGroupName)
 
-                print(layerGroup)
-                print("layer added")
-                layerGroup.name = newGroupName
-                print(newGroupName)
-
-            if app == "QGIS" or app == "ArcGIS": check: Callable[[Base], bool] = lambda base: isinstance(base, Layer) or isinstance(base, VectorLayer) or isinstance(base, RasterLayer)
+            if app == "QGIS" or app == "ArcGIS": check: Callable[[Base], bool] = lambda base: isinstance(base, VectorLayer) or isinstance(base, Layer) or isinstance(base, RasterLayer)
             else: check: Callable[[Base], bool] = lambda base: isinstance(base, Base)
-
-            def callback(base: Base) -> bool:
-                print("callback")
-                #print(base)
-                if isinstance(base, Layer) or isinstance(base, VectorLayer) or isinstance(base, RasterLayer):
-                    layer = layerToNative(base, streamBranch, self.speckleInputs.project)
-                    if layer is not None:
-                        print("Layer created: " + layer.name)
-                else:
-                    loopObj(base, "")
-                return True
+            traverseObject(commitObj, callback, check, str(newGroupName))
             
-            def loopObj(base: Base, baseName: str):
-                memberNames = base.get_member_names()
-                for name in memberNames:
-                    if name in ["id", "applicationId", "units", "speckle_type"]: continue
-                    try: loopVal(base[name], baseName + "/" + name) # loop properties not included above
-                    except: pass
-
-            def loopVal(value: Any, name: str): # "name" is the parent object/property/layer name
-                if name.endswith('/displayValue'): return 
-                
-                if isinstance(value, Base): 
-                    try: # dont go through parts of Speckle Geometry object
-                        print("objects to loop through: " + value.speckle_type)
-                        if value.speckle_type.startswith("Objects.Geometry."): pass #.Brep") or value.speckle_type.startswith("Objects.Geometry.Mesh") or value.speckle_type.startswith("Objects.Geometry.Surface") or value.speckle_type.startswith("Objects.Geometry.Extrusion"): pass
-                        else: loopObj(value, name)
-                    except: loopObj(value, name)
-
-                if isinstance(value, List):
-                    for item in value:
-                        loopVal(item, name)
-                        
-                        print(item)
-                        pt = None
-                        if item.speckle_type and item.speckle_type.startswith("Objects.Geometry."): 
-
-                            pt, pl = cadLayerToNative(value, name, streamBranch, self.speckleInputs.project)
-                            if pt is not None: print("Layer group created: " + pt.name())
-                            if pl is not None: print("Layer group created: " + pl.name())
-                            break
-                        
-                        if item.speckle_type and (item.speckle_type.startswith("Objects.BuiltElements.") or item.speckle_type.startswith("Objects.Structural.Geometry")): #  and "Revit" in item.speckle_type
-                            print("__receiving structures__")
-                            msh_bool = bimLayerToNative(value, name, streamBranch, self.speckleInputs.project)
-                            #if msh is not None: print("Layer group created: " + msh.name())
-                            break
-                
-            traverseObject(commitObj, callback, check)
-      
-        except (SpeckleException, GraphQLException) as e:
-            print("Receive failed: " + str(e))
-            arcpy.AddError("Receive failed: " + str(e))
+        except SpeckleException as e:
+            arcpy.AddWarning("Receive failed: "+ e.message)
             return
+
+    def reloadUI(self):
         
-        print("received")
-        #self.updateParameters(parameters, True)
-        #self.refresh(parameters)
+        try:
+            from speckle.ui.project_vars import get_project_streams, get_survey_point, get_project_layer_selection
+        except: 
+            from speckle_toolbox.esri.toolboxes.speckle.ui.project_vars import get_project_streams, get_survey_point, get_project_layer_selection
+
+        self.is_setup = self.check_for_accounts()
+        if self.dockwidget is not None:
+            self.active_stream = None
+            get_project_streams(self)
+            get_survey_point(self)
+            get_project_layer_selection(self)
+
+            self.dockwidget.reloadDialogUI(self)
+
+    def check_for_accounts(self):
+        def go_to_manager():
+            webbrowser.open("https://speckle-releases.netlify.app/")
+        accounts = get_local_accounts()
+        self.accounts = accounts
+        if len(accounts) == 0:
+            arcpy.AddWarning("No accounts were found. Please remember to install the Speckle Manager and setup at least one account")
+            return False
+        for acc in accounts:
+            if acc.isDefault: 
+                self.default_account = acc 
+                break 
+        return True
+
+    def run(self):
+        """Run method that performs all the real work"""
+        try:
+            from speckle.ui.speckle_qgis_dialog import SpeckleGISDialog
+            from speckle.ui.project_vars import get_project_streams, get_survey_point, get_project_layer_selection
+        except: 
+            from speckle_toolbox.esri.toolboxes.speckle.ui.speckle_qgis_dialog import SpeckleGISDialog
+            from speckle_toolbox.esri.toolboxes.speckle.ui.project_vars import get_project_streams, get_survey_point, get_project_layer_selection
+
+        # Create the dialog with elements (after translation) and keep reference
+        # Only create GUI ONCE in callback, so that it will only load when the plugin is started
+        self.is_setup = self.check_for_accounts()
+            
+        if self.pluginIsActive:
+            self.reloadUI()
+        else:
+            self.pluginIsActive = True
+            if self.dockwidget is None:
+                self.dockwidget = SpeckleGISDialog()
+                self.qgis_project.fileNameChanged.connect(self.reloadUI)
+                self.qgis_project.homePathChanged.connect(self.reloadUI)
+
+            get_project_streams(self)
+            get_survey_point(self)
+            get_project_layer_selection(self)
+
+            self.dockwidget.run(self)
+
+            # show the dockwidget
+            #self.iface.addDockWidget(Qt.RightDockWidgetArea, self.dockwidget)
+            self.dockwidget.enableElements(self)
+
+    def onStreamAddButtonClicked(self):
+        self.add_stream_modal = AddStreamModalDialog(None)
+        self.add_stream_modal.handleStreamAdd.connect(self.handleStreamAdd)
+        self.add_stream_modal.show()
+
+    def set_survey_point(self): 
+        try:
+            from speckle.ui.project_vars import set_survey_point
+        except: 
+            from speckle_toolbox.esri.toolboxes.speckle.ui.project_vars import set_survey_point
+        set_survey_point(self)
+
+    def onStreamCreateClicked(self):
+        self.create_stream_modal = CreateStreamModalDialog(None)
+        self.create_stream_modal.handleStreamCreate.connect(self.handleStreamCreate)
+        #self.create_stream_modal.handleCancelStreamCreate.connect(lambda: self.dockwidget.populateProjectStreams(self))
+        self.create_stream_modal.show()
     
+    def handleStreamCreate(self, account, str_name, description, is_public): 
+        #if len(str_name)<3 and len(str_name)!=0: 
+        #    logger.logToUser("Stream Name should be at least 3 characters", Qgis.Warning)
+        new_client = SpeckleClient(
+            account.serverInfo.url,
+            account.serverInfo.url.startswith("https")
+        )
+        new_client.authenticate_with_token(token=account.token)
+
+        str_id = new_client.stream.create(name=str_name, description = description, is_public = is_public) 
+        if isinstance(str_id, GraphQLException) or isinstance(str_id, SpeckleException):
+            arcpy.addWarning(str_id.message)
+            return
+        else:
+            sw = StreamWrapper(account.serverInfo.url + "/streams/" + str_id)
+            self.handleStreamAdd(sw)
+        return 
+
+    def onBranchCreateClicked(self):
+        self.create_stream_modal = CreateBranchModalDialog(None)
+        self.create_stream_modal.handleBranchCreate.connect(self.handleBranchCreate)
+        self.create_stream_modal.show()
     
-#__all__ = ["Toolbox", "Speckle"]
+    def handleBranchCreate(self, br_name, description):
+        #if len(br_name)<3: 
+        #    logger.logToUser("Branch Name should be at least 3 characters", Qgis.Warning)
+        #    return 
+        br_name = br_name.lower()
+        sw: StreamWrapper = self.active_stream[0]
+        account = sw.get_account()
+        new_client = SpeckleClient(
+            account.serverInfo.url,
+            account.serverInfo.url.startswith("https")
+        )
+        new_client.authenticate_with_token(token=account.token)
+        #description = "No description provided"
+        br_id = new_client.branch.create(stream_id = sw.stream_id, name = br_name, description = description) 
+        if isinstance(br_id, GraphQLException):
+            arcpy.addWarning(br_id.message)
+
+        self.active_stream = (sw, tryGetStream(sw))
+        self.current_streams[0] = self.active_stream
+
+        self.dockwidget.populateActiveStreamBranchDropdown(self)
+        self.dockwidget.populateActiveCommitDropdown(self)
+        self.dockwidget.streamBranchDropdown.setCurrentText(br_name) # will be ignored if branch name is not in the list 
+
+        return 
+
+    def handleStreamAdd(self, sw: StreamWrapper):
+        try:
+            from speckle.ui.project_vars import set_project_streams
+        except:
+            from speckle_toolbox.esri.toolboxes.speckle.ui.project_vars import set_project_streams
+           
+        streamExists = 0
+        index = 0
+        try: 
+            stream = tryGetStream(sw)
+            
+            for st in self.current_streams: 
+                if isinstance(stream, Stream) and st[0].stream_id == stream.id: 
+                    streamExists = 1; 
+                    break 
+                index += 1
+        except SpeckleException as e:
+            arcpy.addWarning(e.message)
+            stream = None
+        
+        if streamExists == 0: 
+            self.current_streams.insert(0,(sw, stream))
+        else: 
+            del self.current_streams[index]
+            self.current_streams.insert(0,(sw, stream))
+        try: self.add_stream_modal.handleStreamAdd.disconnect(self.handleStreamAdd)
+        except: pass 
+        set_project_streams(self)
+        self.dockwidget.populateProjectStreams(self)
